try to fix a bug
